feat: address PR commit
